#177 [REFACTOR] : re-packaging
