+++ app/src/main/java/com/example/charo_android/presentation/ui/charo/mypage/SaveFragment.kt
@@ -1,4 +1,4 @@
-package com.example.charo_android.presentation.ui.charo
+package com.example.charo_android.presentation.ui.charo.mypage
 
 import android.annotation.SuppressLint
 import android.content.Intent
@@ -15,12 +15,14 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.activityViewModels
 import com.example.charo_android.R
 import com.example.charo_android.hidden.Hidden
-import com.example.charo_android.databinding.FragmentMyCharoBinding
+import com.example.charo_android.databinding.FragmentSaveBinding
+import com.example.charo_android.presentation.ui.charo.viewmodel.CharoViewModel
+import com.example.charo_android.presentation.ui.charo.adapter.CharoAdapter
 import com.example.charo_android.presentation.ui.detail.DetailActivity
 
-class MyCharoFragment : Fragment() {
-    private val myCharoViewModel: CharoViewModel by activityViewModels()
-    private var _binding: FragmentMyCharoBinding? = null
+class SaveFragment : Fragment() {
+    private val saveViewModel: CharoViewModel by activityViewModels()
+    private var _binding: FragmentSaveBinding? = null
     private val binding get() = _binding!!
     private lateinit var charoAdapter: CharoAdapter
     var itemLastSize = 0
@@ -38,11 +40,10 @@ class MyCharoFragment : Fragment() {
             intent.putExtra("region", it.region)
             startActivity(intent)
         }
-        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_my_charo, container, false)
+        _binding = DataBindingUtil.inflate(inflater, R.layout.fragment_save, container, false)
         val root: View = binding.root
 
-        Log.d("MyCharoFragment", "created!")
-
+        Log.d("SaveFragment", "created!")
         binding.recyclerviewMyCharo.adapter = charoAdapter
         setUpSpinner()
         setupSpinnerHandler()
@@ -55,7 +56,7 @@ class MyCharoFragment : Fragment() {
     override fun onResume() {
         super.onResume()
         charoAdapter.itemList.clear()
-        myCharoViewModel.clearPostData()
+        saveViewModel.clearPostData()
         setUpSpinner()
         setupSpinnerHandler()
         removeLoading()
@@ -91,14 +92,10 @@ class MyCharoFragment : Fragment() {
                             charoAdapter.spinnerPosition = 0
                         }
 
-                        if (myCharoViewModel.writtenLikeData.value == null) {
-                            myCharoViewModel.getInitLikeData()
-                        }
-
-                        myCharoViewModel.writtenLikeData.observe(viewLifecycleOwner, {
-                            if (myCharoViewModel.writtenLikeData.value?.drive != null) {
+                        saveViewModel.savedLikeData.observe(viewLifecycleOwner, {
+                            if (saveViewModel.savedLikeData.value?.drive != null) {
                                 if(charoAdapter.itemList.isEmpty()) {
-                                    charoAdapter.itemList.addAll(myCharoViewModel.writtenLikeData.value?.drive!!)
+                                    charoAdapter.itemList.addAll(saveViewModel.savedLikeData.value?.drive!!)
                                     charoAdapter.notifyDataSetChanged()
                                     itemLastSize = charoAdapter.itemList.size
                                 }
@@ -110,14 +107,14 @@ class MyCharoFragment : Fragment() {
                             charoAdapter.spinnerPosition = 1
                         }
 
-                        if (myCharoViewModel.writtenNewData.value == null) {
-                            myCharoViewModel.getInitNewData()
+                        if (saveViewModel.writtenNewData.value == null) {
+                            saveViewModel.getInitNewData()
                         }
 
-                        myCharoViewModel.writtenNewData.observe(viewLifecycleOwner, {
-                            if (myCharoViewModel.writtenNewData.value?.drive != null) {
+                        saveViewModel.savedNewData.observe(viewLifecycleOwner, {
+                            if (saveViewModel.savedNewData.value?.drive != null) {
                                 if(charoAdapter.itemList.isEmpty()) {
-                                    charoAdapter.itemList.addAll(myCharoViewModel.writtenNewData.value?.drive!!)
+                                    charoAdapter.itemList.addAll(saveViewModel.savedNewData.value?.drive!!)
                                     charoAdapter.notifyDataSetChanged()
                                     itemLastSize = charoAdapter.itemList.size
                                 }
@@ -139,18 +136,17 @@ class MyCharoFragment : Fragment() {
                 if ((scrollY >= (v.getChildAt(v.childCount - 1).measuredHeight - v.measuredHeight)) &&
                     scrollY > oldScrollY
                 ) {
-                    Log.d("무한스크롤 최하단 도달", "도달완료")
-                    if (myCharoViewModel.isServerConnection.value == false && charoAdapter.itemList.isNotEmpty()) {
+                    Log.d("무한스크롤 최하단 도달", "도달완")
+                    if (saveViewModel.isServerConnection.value == false && charoAdapter.itemList.isNotEmpty()) {
                         charoAdapter.addLoading()
                         when (spinnerPosition) {
                             0 -> {
-                                myCharoViewModel.getMoreWrittenLikeData()
+                                saveViewModel.getMoreSavedLikeData()
                             }
                             else -> {
-                                myCharoViewModel.getMoreWrittenNewData()
+                                saveViewModel.getMoreSavedNewData()
                             }
                         }
-                        charoAdapter.removeLoading()
                     }
                 }
             }
@@ -158,7 +154,7 @@ class MyCharoFragment : Fragment() {
     }
 
     private fun removeLoading() {
-        myCharoViewModel.isServerConnection.observe(viewLifecycleOwner, {
+        saveViewModel.isServerConnection.observe(viewLifecycleOwner, {
             if (charoAdapter.itemList.isNotEmpty()) {
                 charoAdapter.removeLoading()
             }
@@ -166,24 +162,24 @@ class MyCharoFragment : Fragment() {
     }
 
     private fun getMoreLikeData() {
-        myCharoViewModel.writtenMoreLikeData.observe(viewLifecycleOwner, {
-            if (myCharoViewModel.writtenLikeData.value?.drive != null && myCharoViewModel.writtenMoreLikeData.value?.drive != null) {
-                charoAdapter.itemList.addAll(myCharoViewModel.writtenMoreLikeData.value?.drive!!)
+        saveViewModel.savedMoreLikeData.observe(viewLifecycleOwner, {
+            if (saveViewModel.savedLikeData.value?.drive != null && saveViewModel.savedMoreLikeData.value?.drive != null) {
+                charoAdapter.itemList.addAll(saveViewModel.savedMoreLikeData.value?.drive!!)
                 charoAdapter.notifyItemRangeInserted(
                     itemLastSize,
-                    myCharoViewModel.writtenMoreLikeData.value?.drive!!.size
+                    saveViewModel.savedMoreLikeData.value?.drive!!.size
                 )
             }
         })
     }
 
     private fun getMoreNewData() {
-        myCharoViewModel.writtenMoreNewData.observe(viewLifecycleOwner, {
-            if (myCharoViewModel.writtenNewData.value?.drive != null && myCharoViewModel.writtenMoreNewData.value?.drive != null) {
-                charoAdapter.itemList.addAll(myCharoViewModel.writtenMoreNewData.value?.drive!!)
+        saveViewModel.savedMoreNewData.observe(viewLifecycleOwner, {
+            if (saveViewModel.savedNewData.value?.drive != null && saveViewModel.savedMoreLikeData.value?.drive != null) {
+                charoAdapter.itemList.addAll(saveViewModel.savedMoreNewData.value?.drive!!)
                 charoAdapter.notifyItemRangeInserted(
                     itemLastSize,
-                    myCharoViewModel.writtenMoreNewData.value?.drive!!.size
+                    saveViewModel.savedMoreNewData.value?.drive!!.size
                 )
             }
         })
