+++ app/src/main/java/com/example/charo_android/presentation/ui/charo/otherpage/OtherCharoFragment.kt
@@ -1,4 +1,4 @@
-package com.example.charo_android.presentation.ui.charo
+package com.example.charo_android.presentation.ui.charo.otherpage
 
 import android.content.Intent
 import android.os.Bundle
@@ -10,7 +10,8 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.activityViewModels
 import com.example.charo_android.R
 import com.example.charo_android.databinding.FragmentOtherCharoBinding
-import com.example.charo_android.hidden.Hidden
+import com.example.charo_android.presentation.ui.charo.viewmodel.CharoViewModel
+import com.example.charo_android.presentation.ui.charo.follow.CharoListActivity
 
 class OtherCharoFragment : Fragment() {
     private val charoViewModel: CharoViewModel by activityViewModels()
