+++ app/src/main/res/layout/activity_charo_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.charo.CharoListActivity">
+    tools:context=".presentation.ui.charo.follow.CharoListActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/charo_list_fragment"
