+++ app/src/main/res/layout/fragment_charo_list.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="charoListData"
-            type="com.example.charo_android.presentation.ui.charo.CharoViewModel" />
+            type="com.example.charo_android.presentation.ui.charo.viewmodel.CharoViewModel" />
 
         <import type="android.view.View" />
     </data>
@@ -15,7 +15,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.charo.CharoListFragment">
+        tools:context=".presentation.ui.charo.follow.CharoListFragment">
 
         <ImageView
             android:id="@+id/img_charo_list_back"
