+++ app/src/main/res/layout/fragment_my_charo.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="myPageData"
-            type="com.example.charo_android.presentation.ui.charo.CharoViewModel" />
+            type="com.example.charo_android.presentation.ui.charo.viewmodel.CharoViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
