+++ app/src/main/res/layout/fragment_other_write_charo.xml
@@ -10,7 +10,7 @@
 
         <variable
             name="otherPageData"
-            type="com.example.charo_android.presentation.ui.charo.CharoViewModel" />
+            type="com.example.charo_android.presentation.ui.charo.viewmodel.CharoViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
